% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProfiles.R
\name{readProfiles}
\alias{readProfiles}
\title{Read a set of profiles from a file}
\usage{
readProfiles(fileName, freqs, delim = "\\t", header = FALSE, id = 1)
}
\arguments{
\item{fileName}{a path to the profile file.}

\item{delim}{a character that delimits the fields in the profile file.}

\item{header}{a boolean witch is \code{TRUE} if the profile file has a column header line.}

\item{Freqs}{A list containing two lists labelled loci and freqs. The second
list is a list of vectors containing the allele frequencies of each allele
at each locus in the multiplex.}
}
\value{
matrix of profiles.
}
\description{
Reads a set of profiles from a file
}
\details{
The alleles are recorded integers rather than the STR repeat numbers. This
speeds up computation immensely when calculating any of the LRs or IBS.
}
\examples{

data(fbiCaucs)

}
\author{
James M. Curran
}
