% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relSim-package.R
\docType{package}
\name{relSim}
\alias{relSim}
\alias{relSim-package}
\title{Relative Simulator}
\description{
relSim
}
\details{
A set of tools to explore the behaviour statistics used for forensic DNA interpretation when close relatives are involved. The package also offers some useful tools for exploring other forensic DNA situations.
}
\author{
James Curran
}
