% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.population.R
\name{print.population}
\alias{print.population}
\title{Print summary details of a substructed population}
\usage{
\method{print}{population}(x, ...)
}
\arguments{
\item{x}{The population object to be printed}

\item{\dots}{Ignored - really should be passed to print, but given cat is
actually called they are ignored}
}
\description{
Nicely prints summary information about a substructured population created
using \code{breedFst}
}
\examples{

data(fbiCaucs)
p = breedFst(fbiCaucs)
print(p)

}
\seealso{
breedFst
}
\author{
James M. Curran
}
